DIL-2333: Cleaning up the netID display code
- app/assets/javascripts/dil_collections.js

Remove whitespace. Yes, it matters.

- app/controllers/dil_collections_controller.rb

Moving the User.admin? checks and JSON modification here. Get the JSON from the
model as a string and re-JSONize. Iterate over the collection and add an
"owner" element if the current_user is an admin. If the "owner" element is
empty it just returns nil.

- app/models/dil_collection.rb

Moving the User.admin? checks and JSON modification out of the model because we
can't access the current_user there.