UI fixes for public users
app/views/multiresimages/_crop.html.erb

Modify the call to JavaScript, adding logic to account for the user's
sign-in status. Call the user_signed_in? method: if the user is signed
in, display the crop tool; if they're not, don't.

Add logic to the view to account for the user's sign-in status. Call
the user_signed_in? method: if the user is signed in, display the
"Add to Image Group" button; if they're not, don't.